Add delete button and add code editor to editing code snippet ui
Adds a delete button to each code snippet. Add code editor to the
code snippet editor dialog. Move CodeSnippetDisplay into its own
file.
