+++ packages/code-snippet/src/CodeSnippetWidget.tsx
@@ -17,24 +17,18 @@
 import '../style/index.css';
 
 import { SubmissionHandler } from '@elyra/application';
-import { ExpandableComponent, addSnippetIcon } from '@elyra/ui-components';
+import { addSnippetIcon } from '@elyra/ui-components';
 
 import {
   ReactWidget,
   UseSignal,
-  Clipboard,
   Dialog,
   showDialog
 } from '@jupyterlab/apputils';
-import { CodeCell, MarkdownCell } from '@jupyterlab/cells';
 import { CodeEditor } from '@jupyterlab/codeeditor';
-import { PathExt } from '@jupyterlab/coreutils';
+// import { CodeMirrorEditorFactory } from '@jupyterlab/codemirror';
 import { IDocumentManager } from '@jupyterlab/docmanager';
-import { DocumentWidget } from '@jupyterlab/docregistry';
-import { FileEditor } from '@jupyterlab/fileeditor';
-import { Notebook, NotebookPanel } from '@jupyterlab/notebook';
 import { IRenderMimeRegistry } from '@jupyterlab/rendermime';
-import { copyIcon, addIcon, editIcon } from '@jupyterlab/ui-components';
 import { Message } from '@lumino/messaging';
 import { Signal } from '@lumino/signaling';
 import { Widget } from '@lumino/widgets';
@@ -42,6 +36,7 @@ import { Widget } from '@lumino/widgets';
 import React from 'react';
 
 import { CodeSnippetManager, ICodeSnippet } from './CodeSnippet';
+import { CodeSnippetDisplay } from './CodeSnippetDisplay';
 import { EditorDialog } from './EditorDialog';
 
 /**
@@ -49,181 +44,12 @@ import { EditorDialog } from './EditorDialog';
  */
 const CODE_SNIPPETS_CLASS = 'elyra-CodeSnippets';
 const CODE_SNIPPETS_HEADER_CLASS = 'elyra-codeSnippetsHeader';
-const CODE_SNIPPET_ITEM = 'elyra-codeSnippet-item';
 const CODE_SNIPPETS_HEADER_BUTTON_CLASS = 'elyra-codeSnippetHeader-button';
 
 /**
- * CodeSnippetDisplay props.
+ * API endpoint for code snippets
  */
-interface ICodeSnippetDisplayProps {
-  codeSnippets: ICodeSnippet[];
-  getCurrentWidget: () => Widget;
-  editCodeSnippet: any;
-}
-
-/**
- * A React Component for code-snippets display list.
- */
-class CodeSnippetDisplay extends React.Component<ICodeSnippetDisplayProps> {
-  // TODO: Use code mirror to display code
-
-  // Handle code snippet insert into an editor
-  private insertCodeSnippet = async (snippet: ICodeSnippet): Promise<void> => {
-    const widget: Widget = this.props.getCurrentWidget();
-    const snippetStr: string = snippet.code.join('\n');
-
-    if (
-      widget instanceof DocumentWidget &&
-      (widget as DocumentWidget).content instanceof FileEditor
-    ) {
-      const documentWidget = widget as DocumentWidget;
-      const fileEditor = (documentWidget.content as FileEditor).editor;
-      const markdownRegex = /^\.(md|mkdn?|mdown|markdown)$/;
-      if (PathExt.extname(widget.context.path).match(markdownRegex) !== null) {
-        // Wrap snippet into a code block when inserting it into a markdown file
-        fileEditor.replaceSelection(
-          '```' + snippet.language + '\n' + snippetStr + '\n```'
-        );
-      } else if (widget.constructor.name == 'PythonFileEditor') {
-        this.verifyLanguageAndInsert(snippet, 'python', fileEditor);
-      } else {
-        fileEditor.replaceSelection(snippetStr);
-      }
-    } else if (widget instanceof NotebookPanel) {
-      const notebookWidget = widget as NotebookPanel;
-      const notebookCell = (notebookWidget.content as Notebook).activeCell;
-      const notebookCellEditor = notebookCell.editor;
-
-      if (notebookCell instanceof CodeCell) {
-        const kernelInfo = await notebookWidget.sessionContext.session?.kernel
-          ?.info;
-        const kernelLanguage: string = kernelInfo?.language_info.name || '';
-        this.verifyLanguageAndInsert(
-          snippet,
-          kernelLanguage,
-          notebookCellEditor
-        );
-      } else if (notebookCell instanceof MarkdownCell) {
-        // Wrap snippet into a code block when inserting it into a markdown cell
-        notebookCellEditor.replaceSelection(
-          '```' + snippet.language + '\n' + snippetStr + '\n```'
-        );
-      } else {
-        notebookCellEditor.replaceSelection(snippetStr);
-      }
-    } else {
-      this.showErrDialog('Code snippet insert failed: Unsupported widget');
-    }
-  };
-
-  // Handle language compatibility between code snippet and editor
-  private verifyLanguageAndInsert = async (
-    snippet: ICodeSnippet,
-    editorLanguage: string,
-    editor: CodeEditor.IEditor
-  ): Promise<void> => {
-    const snippetStr: string = snippet.code.join('\n');
-    if (
-      editorLanguage &&
-      snippet.language.toLowerCase() !== editorLanguage.toLowerCase()
-    ) {
-      const result = await this.showWarnDialog(
-        editorLanguage,
-        snippet.displayName
-      );
-      if (result.button.accept) {
-        editor.replaceSelection(snippetStr);
-      }
-    } else {
-      // Language match or editorLanguage is unavailable
-      editor.replaceSelection(snippetStr);
-    }
-  };
-
-  // Display warning dialog when inserting a code snippet incompatible with editor's language
-  private showWarnDialog = async (
-    editorLanguage: string,
-    snippetName: string
-  ): Promise<Dialog.IResult<string>> => {
-    return showDialog({
-      title: 'Warning',
-      body:
-        'Code snippet "' +
-        snippetName +
-        '" is incompatible with ' +
-        editorLanguage +
-        '. Continue?',
-      buttons: [Dialog.cancelButton(), Dialog.okButton()]
-    });
-  };
-
-  // Display error dialog when inserting a code snippet into unsupported widget (i.e. not an editor)
-  private showErrDialog = (errMsg: string): Promise<Dialog.IResult<string>> => {
-    return showDialog({
-      title: 'Error',
-      body: errMsg,
-      buttons: [Dialog.okButton()]
-    });
-  };
-
-  // Render display of code snippet list
-  private renderCodeSnippet = (codeSnippet: ICodeSnippet): JSX.Element => {
-    const displayName =
-      '[' + codeSnippet.language + '] ' + codeSnippet.displayName;
-
-    const actionButtons = [
-      {
-        title: 'Copy',
-        icon: copyIcon,
-        onClick: (): void => {
-          Clipboard.copyToSystem(codeSnippet.code.join('\n'));
-        }
-      },
-      {
-        title: 'Insert',
-        icon: addIcon,
-        onClick: (): void => {
-          this.insertCodeSnippet(codeSnippet);
-        }
-      },
-      {
-        title: 'Edit',
-        icon: editIcon,
-        onClick: (): void => {
-          this.props.editCodeSnippet(
-            codeSnippet.displayName,
-            codeSnippet.description,
-            codeSnippet.language,
-            codeSnippet.code.join('\n'),
-            false
-          );
-        }
-      }
-    ];
-
-    return (
-      <div key={codeSnippet.name} className={CODE_SNIPPET_ITEM}>
-        <ExpandableComponent
-          displayName={displayName}
-          tooltip={codeSnippet.description}
-          actionButtons={actionButtons}
-        >
-          <textarea defaultValue={codeSnippet.code.join('\n')}></textarea>
-        </ExpandableComponent>
-      </div>
-    );
-  };
-
-  render(): React.ReactElement {
-    return (
-      <div>
-        <div id="codeSnippets">
-          <div>{this.props.codeSnippets.map(this.renderCodeSnippet)}</div>
-        </div>
-      </div>
-    );
-  }
-}
+export const CODE_SNIPPET_ENDPOINT = 'api/metadata/code-snippets';
 
 /**
  * A widget for Code Snippets.
@@ -232,12 +58,18 @@ export class CodeSnippetWidget extends ReactWidget {
   codeSnippetManager: CodeSnippetManager;
   renderCodeSnippetsSignal: Signal<this, ICodeSnippet[]>;
   getCurrentWidget: () => Widget;
+  editorFactory: CodeEditor.Factory;
 
-  constructor(getCurrentWidget: () => Widget) {
+  constructor(
+    getCurrentWidget: () => Widget,
+    editorFactory: CodeEditor.Factory
+  ) {
     super();
     this.getCurrentWidget = getCurrentWidget;
     this.codeSnippetManager = new CodeSnippetManager();
     this.renderCodeSnippetsSignal = new Signal<this, ICodeSnippet[]>(this);
+    this.editCodeSnippet = this.editCodeSnippet.bind(this);
+    this.editorFactory = editorFactory;
   }
 
   // Request code snippets from server
@@ -267,14 +99,14 @@ export class CodeSnippetWidget extends ReactWidget {
     code: string,
     newFile: boolean
   ): Promise<void> {
-    const codeSnippetEndpoint = 'api/metadata/code-snippets';
     showDialog({
       title: 'Edit snippet',
       body: new EditorDialog({
         display_name: displayName,
         description: description,
         language: language,
-        code: code
+        code: code,
+        editorFactory: this.editorFactory
       }),
       buttons: [Dialog.cancelButton(), Dialog.okButton()]
     }).then((result: any) => {
@@ -297,7 +129,7 @@ export class CodeSnippetWidget extends ReactWidget {
 
       if (newFile) {
         SubmissionHandler.makePostRequest(
-          codeSnippetEndpoint,
+          CODE_SNIPPET_ENDPOINT,
           JSON.stringify(newSnippet),
           'code snippets',
           (response: any) => {
@@ -306,7 +138,7 @@ export class CodeSnippetWidget extends ReactWidget {
         );
       } else {
         SubmissionHandler.makeServerRequest(
-          codeSnippetEndpoint + '/' + newSnippet.name,
+          CODE_SNIPPET_ENDPOINT + '/' + newSnippet.name,
           { method: 'PUT', body: newSnippetString },
           'code snippets',
           (response: any) => {
@@ -339,6 +171,7 @@ export class CodeSnippetWidget extends ReactWidget {
               codeSnippets={codeSnippets}
               editCodeSnippet={this.editCodeSnippet}
               getCurrentWidget={this.getCurrentWidget}
+              updateSnippets={this.updateSnippets}
             />
           )}
         </UseSignal>
