+++ packages/code-snippet/src/index.ts
@@ -24,6 +24,8 @@ import {
   ILayoutRestorer
 } from '@jupyterlab/application';
 import { ICommandPalette } from '@jupyterlab/apputils';
+import { IEditorServices } from '@jupyterlab/codeeditor';
+// import { CodeMirrorEditorFactory } from '@jupyterlab/codemirror';
 
 import { Widget } from '@lumino/widgets';
 
@@ -37,11 +39,12 @@ const CODE_SNIPPET_EXTENSION_ID = 'elyra-code-snippet-extension';
 export const code_snippet_extension: JupyterFrontEndPlugin<void> = {
   id: CODE_SNIPPET_EXTENSION_ID,
   autoStart: true,
-  requires: [ICommandPalette, ILayoutRestorer],
+  requires: [ICommandPalette, ILayoutRestorer, IEditorServices],
   activate: (
     app: JupyterFrontEnd,
     palette: ICommandPalette,
-    restorer: ILayoutRestorer
+    restorer: ILayoutRestorer,
+    editorServices: IEditorServices
   ) => {
     console.log('Elyra - code-snippet extension is activated!');
 
@@ -49,7 +52,10 @@ export const code_snippet_extension: JupyterFrontEndPlugin<void> = {
       return app.shell.currentWidget;
     };
 
-    const codeSnippetWidget = new CodeSnippetWidget(getCurrentWidget);
+    const codeSnippetWidget = new CodeSnippetWidget(
+      getCurrentWidget,
+      editorServices.factoryService.newInlineEditor
+    );
     codeSnippetWidget.id = CODE_SNIPPET_EXTENSION_ID;
     codeSnippetWidget.title.icon = codeSnippetIcon;
     codeSnippetWidget.title.caption = 'Code Snippet';
